outline new editor ui
