+++ app/components/Editor/Sidebar/index.tsx
@@ -1,9 +1,8 @@
-import { Box, ThemeContext } from "grommet";
+import { Box } from "grommet";
 import { useContext, useState } from "react";
 import { Resizable, ResizeCallback } from "re-resizable";
 
 import { NavigationOption } from "../../../lib/types";
-import { theme } from "../../../theme/theme-new";
 import CodeEditor from "./CodeEditor";
 import HelpersEditor from "./HelpersEditor";
 import Navigation from "./Navigation";
@@ -33,25 +32,23 @@ export default function Sidebar(): JSX.Element {
   };
 
   return (
-    <ThemeContext.Extend value={theme}>
-      <Resizable
-        defaultSize={{
-          height: "100%",
-          width: editorSidebarWidth,
-        }}
-        enable={enable}
-        maxWidth="50%"
-        minWidth={480}
-        onResizeStop={handleResizeStop}
-      >
-        <Box background="gray10" flex={false} height="full">
-          <Navigation selected={selected} setSelected={setSelected} />
-          {selected === "code" && <CodeEditor />}
-          {selected === "helpers" && <HelpersEditor />}
-          <RunLogs isVisible={selected === "logs"} />
-          <Buttons />
-        </Box>
-      </Resizable>
-    </ThemeContext.Extend>
+    <Resizable
+      defaultSize={{
+        height: "100%",
+        width: editorSidebarWidth,
+      }}
+      enable={enable}
+      maxWidth="50%"
+      minWidth={480}
+      onResizeStop={handleResizeStop}
+    >
+      <Box background="gray10" flex={false} height="full">
+        <Navigation selected={selected} setSelected={setSelected} />
+        {selected === "code" && <CodeEditor />}
+        {selected === "helpers" && <HelpersEditor />}
+        <RunLogs isVisible={selected === "logs"} />
+        <Buttons />
+      </Box>
+    </Resizable>
   );
 }
